move http-bridge specific integration scenario in the 'http-bridge' package
