+++ lib/core/test/integration/Test/Integration/Scenario/API/Transactions.hs
@@ -55,7 +55,6 @@ import Test.Integration.Framework.DSL
 import Test.Integration.Framework.TestData
     ( arabicWalletName
     , errMsg403Fee
-    , errMsg403InvalidTransaction
     , errMsg403NotEnoughMoney
     , errMsg403UTxO
     , errMsg403WrongPass
@@ -705,56 +704,3 @@ spec = do
             r <- request @(ApiTransaction t) ctx (postTxEp w)
                     Default payload
             expectResponseCode @IO HTTP.status400 r
-
-    it "TRANS_CREATE_09 - 0 amount transaction is forbidden on single output tx" $ \ctx -> do
-        wSrc <- fixtureWallet ctx
-        wDest <- emptyWallet ctx
-        addr:_ <- listAddresses ctx wDest
-
-        let destination = addr ^. #id
-        let payload = Json [json|{
-                "payments": [{
-                    "address": #{destination},
-                    "amount": {
-                        "quantity": 0,
-                        "unit": "lovelace"
-                    }
-                }],
-                "passphrase": "cardano-wallet"
-            }|]
-        r <- request @(ApiTransaction t) ctx (postTxEp wSrc) Default payload
-        verify r
-            [ expectResponseCode HTTP.status403
-            , expectErrorMessage errMsg403InvalidTransaction
-            ]
-
-    it "TRANS_CREATE_09 - 0 amount transaction is forbidden on multi-output tx" $ \ctx -> do
-        wSrc <- fixtureWallet ctx
-        wDest <- emptyWallet ctx
-        addrs <- listAddresses ctx wDest
-
-        let destination1 = (addrs !! 1) ^. #id
-        let destination2 = (addrs !! 2) ^. #id
-        let payload = Json [json|{
-                "payments": [{
-                    "address": #{destination1},
-                    "amount": {
-                        "quantity": 0,
-                        "unit": "lovelace"
-                    }
-                },
-                {
-                    "address": #{destination2},
-                    "amount": {
-                        "quantity": 23,
-                        "unit": "lovelace"
-                    }
-                }],
-                "passphrase": "cardano-wallet"
-            }|]
-
-        r <- request @(ApiTransaction t) ctx (postTxEp wSrc) Default payload
-        verify r
-            [ expectResponseCode HTTP.status403
-            , expectErrorMessage errMsg403InvalidTransaction
-            ]
