--- lib/http-bridge/test/integration/Test
@@ -1,1 +1,0 @@
-../../../core/test/integration/Test
