+++ lib/http-bridge/test/integration/Test/Integration/Faucet.hs
@@ -1,0 +1,1 @@
+../../../../../core/test/integration/Test/Integration/Faucet.hs
